+++ src/bulk_extractor_scanners.h
@@ -53,6 +53,7 @@ SCANNER(outlook)
 SCANNER(pdf)
 SCANNER(rar)
 SCANNER(rtti)
+SCANNER(secrets)
 SCANNER(sqlite)
 SCANNER(utmp)        // scanner provided by 4n6ist:
 SCANNER(vcard)